refactor: keep the core helpers in one _utils module
Merge _event_time and _loading back into core/_utils. A general _utils
holds a package's assorted small helpers, and a helper earns its own
module only when it takes on a distinct role worth a name. The public
surface is unchanged: core re-exports format_event_time, parse_event_time
and load_object as before.
