+++ src/sportsbet/core/__init__.py
@@ -1,8 +1,6 @@
 """Provide the shared types, constants and building primitives."""
 
 from ._errors import BuildError
-from ._event_time import format_event_time, parse_event_time
-from ._loading import load_object
 from ._params import (
     ALIASES,
     DATE_COLS,
@@ -28,6 +26,7 @@ from ._types import (
     Schema,
     TrainData,
 )
+from ._utils import format_event_time, load_object, parse_event_time
 
 __all__ = [
     'ALIASES',
